assignment-2-784478/code/batchingest-manager/requirements.txt
flask
apscheduler
pymongo
pyyaml
watchdog
